pycam/requirements.txt
PyOpenGL
PyYAML
svg.path
